correct type

+++ lmu_type.py
@@ -228,7 +228,7 @@ class LMUVehicleScoring(_NOINIT):
     mSteamID: int
     mVehFilename: bytes
     mAttackMode: int
-    mFuelFraction: float
+    mFuelFraction: int
     mDRSState: bool
     mExpansion: tuple[int, ...]
